@@ -1,76 +1,78 @@
 {
-  "name": "react-openlayers",
-  "version": "0.2.2",
-  "description": "OpenLayer React Components",
-  "license": "MIT",
-  "main": "dist/react-openlayers.umd.js",
-  "module": "dist/index.js",
-  "typings": "dist/index.d.ts",
-  "scripts": {
-    "start": "cross-env NODE_ENV=dev webpack-dev-server --quiet --port 9001 --content-base app --config app/webpack.config --open",
-    "clean": "rimraf dist",
-    "build": "npm-run-all --serial clean build:umd build:app",
-    "build:umd": "cross-env NODE_ENV=prod webpack",
-    "build:app": "cross-env NODE_ENV=prod webpack --config app/webpack.config",
-    "upgrade": "npm-check-updates -a/--upgradeAll && npm i"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/allenhwkim/react-openlayers.git"
-  },
-  "author": "Allen Kim<allenhwkim@gmail.com>",
-  "bugs": {
-    "url": "https://github.com/allenhwkim/react-openlayers/issues"
-  },
-  "homepage": "https://github.com/allenhwkim/react-openlayers#readme",
-  "keywords": [
-    "openlayer",
-    "typescript",
-    "react",
-    "map"
-  ],
-  "devDependencies": {
-    "@types/node": "^6.0.31",
-    "@types/openlayers": "^3.20.4",
-    "@types/react": "^15.0.11",
-    "@types/react-dom": "^0.14.23",
-    "@types/whatwg-fetch": "0.0.33",
-    "awesome-typescript-loader": "^3.0.8",
-    "core-js": "^2.4.1",
-    "css-loader": "^0.26.2",
-    "forever": "^0.15.3",
-    "fs-extra": "^2.0.0",
-    "geo-coder": "^0.2.1",
-    "html-loader": "^0.4.5",
-    "node-sass": "^4.5.0",
-    "npm-run-all": "^3.1.2",
-    "openlayers": "^4.0.1",
-    "raw-loader": "^0.5.1",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "react-router": "^3.0.2",
-    "resolve-url-loader": "^2.0.1",
-    "rimraf": "^2.6.0",
-    "rxjs": "^5.0.0-rc.1",
-    "sass-loader": "^6.0.2",
-    "source-map-loader": "^0.1.6",
-    "strip-loader": "^0.1.2",
-    "style-loader": "^0.13.2",
-    "ts-loader": "^1.3.3",
-    "typescript": "2.1.6",
-    "url-loader": "^0.5.8",
-    "webpack": "^2.2.1",
-    "webpack-dev-server": "^2.4.1",
-    "webtest": "^0.3.6",
-    "zone.js": "^0.6.21"
-  },
-  "dependencies": {
-    "@types/googlemaps": "^3.26.4",
-    "cross-env": "^4.0.0",
-    "google-maps": "^3.2.1",
-    "ol-mapbox-style": "^2.1.1",
-    "qwest": "^4.4.5",
-    "react-async-loader": "^0.1.2",
-    "react-async-script-loader": "^0.2.2"
-  }
+    "name": "react-openlayers",
+    "version": "0.2.6",
+    "description": "OpenLayer React Components",
+    "license": "MIT",
+    "main": "dist/index.js",
+    "module": "dist/index.js",
+    "types": "dist/index.d.ts",
+    "files": [
+        "dist",
+        "src"
+    ],
+    "scripts": {
+        "start": "cross-env NODE_ENV=development webpack-dev-server --port 9001 --content-base app --config app/webpack.config --open",
+        "clean": "rimraf dist",
+        "build": "npm-run-all --serial clean build:tsc copy:static",
+        "build:tsc": "tsc",
+        "build:umd": "cross-env NODE_ENV=production webpack",
+        "build:app": "cross-env NODE_ENV=production webpack --config app/webpack.config",
+        "types": "dts-generator --name react-openlayers --project . --out dist/index.d.ts",
+        "copy:static": "cp src/map.css dist/map.css && cp src/custom/popup.css dist/custom/popup.css",
+        "packy": "npm pack && mv react-openlayers-0.2.6.tgz /Users/keithmoss/Downloads/react-openlayers-0.2.6.tgz",
+        "wml:add": "wml add ./ /Users/keithmoss/Documents/Work/GitHub/ealgis-yarn/frontend/node_modules/react-openlayers",
+        "wml:start": "wml start",
+        "upgrade": "npm-check-updates -a/--upgradeAll && npm i"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/allenhwkim/react-openlayers.git"
+    },
+    "author": "Allen Kim<allenhwkim@gmail.com>",
+    "bugs": {
+        "url": "https://github.com/allenhwkim/react-openlayers/issues"
+    },
+    "homepage": "https://github.com/allenhwkim/react-openlayers#readme",
+    "keywords": [
+        "openlayer",
+        "typescript",
+        "react",
+        "map"
+    ],
+    "devDependencies": {
+        "@types/googlemaps": "^3.26.4",
+        "@types/ol": "^4.6.1",
+        "@types/react": "^16.3.11",
+        "@types/react-dom": "^16.0.4",
+        "awesome-typescript-loader": "^5.0.0",
+        "babel-cli": "^6.26.0",
+        "babel-preset-es2015": "^6.24.1",
+        "cross-env": "^4.0.0",
+        "css-loader": "^0.26.2",
+        "dts-generator": "^2.1.0",
+        "file-loader": "^1.1.11",
+        "geo-coder": "^0.2.1",
+        "google-maps": "^3.2.1",
+        "html-loader": "^0.4.5",
+        "node-sass": "^4.5.0",
+        "npm-run-all": "^3.1.2",
+        "ol": "^4.6.5",
+        "ol-mapbox-style": "^2.11.2",
+        "qwest": "^4.4.5",
+        "react": "^16.3.0",
+        "react-dom": "^16.3.0",
+        "react-router": "^3.0.2",
+        "rimraf": "^2.6.0",
+        "style-loader": "^0.13.2",
+        "ts-loader": "^4.1.0",
+        "typescript": "^2.8.1",
+        "webpack": "^4.4.1",
+        "webpack-cli": "^2.0.13",
+        "webpack-dev-server": "^3.1.1"
+    },
+    "peerDependencies": {
+        "ol": "^4.6.5",
+        "react": "^16.3.0",
+        "react-dom": "^16.3.0"
+    }
 }
