@@ -1,15 +1,16 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
-import * as ol from 'openlayers';
+import olSourceOSM from "ol/source/osm";
+import olSourceVector from "ol/source/vector";
 import {
   interaction, layer, custom, control, //name spaces
   Interactions, Overlays, Controls,     //group
   Map, Layers, Overlay, Util    //objects
 } from "react-openlayers";
 
-var rasterTile = new ol.source.OSM();
+var rasterTile = new olSourceOSM();
 
-var vectorSource = new ol.source.Vector({wrapX: false});
+var vectorSource = new olSourceVector({wrapX: false});
 
 export class Draw extends React.Component<any, any> {
   constructor(props) {
