@@ -1,6 +1,6 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
-import * as ol from 'openlayers';
+import olSourceStamen from 'ol/source/stamen';
 import {
   interaction, layer, custom, control, //name spaces
   Interactions, Overlays, Controls,     //group
@@ -17,14 +17,14 @@ export class Tile extends React.Component<any,any> {
       <div>
         <Map>
           <Layers>
-            <layer.Tile source={new ol.source.Stamen({ layer: 'watercolor' })}/>
+            <layer.Tile source={new olSourceStamen({ layer: 'watercolor' })}/>
           </Layers>
         </Map>
         <a href="https://github.com/allenhwkim/react-openlayers/blob/master/app/layers/tile.tsx">Source Code</a>
         <pre>{`
         <Map>
           <Layers>
-            <layer.Tile source={new ol.source.Stamen({ layer: 'watercolor' })}/>
+            <layer.Tile source={new olSourceStamen({ layer: 'watercolor' })}/>
           </Layers>
         </Map>
         `}</pre>
