@@ -1,14 +1,16 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
-import * as ol from 'openlayers';
+import olFeature from "ol/feature";
+import olSoureVector from "ol/source/vector";
+import olGeomPoint from "ol/geom/point";
 import {
   interaction, layer, custom, control, //name spaces
   Interactions, Overlays, Controls,     //group
   Map, Layers, Overlay, Util    //objects
 } from "react-openlayers";
 
-var iconFeature = new ol.Feature(new ol.geom.Point([0, 0]));
-var source = new ol.source.Vector({features: [iconFeature]});
+var iconFeature = new olFeature(new olGeomPoint([0, 0]));
+var source = new olSoureVector({features: [iconFeature]});
 var marker = new custom.style.MarkerStyle(
   'https://openlayers.org/en/v4.0.1/examples/data/icon.png'
 );
