@@ -1,6 +1,7 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
-import * as ol from 'openlayers';
+import olProj from 'ol/proj';
+import olSourceStamen from "ol/source/stamen";
 import {
   interaction, layer, custom, control, //name spaces
   Interactions, Overlays, Controls,     //group
@@ -18,7 +19,7 @@ export class AppOverlay extends React.Component<any,any> {
 
   showPopup = (evt) => {
     this.overlayComp.overlay.setPosition(evt.coordinate);
-    var lonlat = ol.proj.transform(evt.coordinate, 'EPSG:3857', 'EPSG:4326');
+    var lonlat = olProj.transform(evt.coordinate, 'EPSG:3857', 'EPSG:4326');
 
     this.popupComp.setContents(
       `<p>You clicked here:</p><code> ${lonlat[0]}, ${lonlat[1]}</code>`
@@ -31,7 +32,7 @@ export class AppOverlay extends React.Component<any,any> {
       <div>
         <Map onClick={this.showPopup}>
           <Layers>
-            <layer.Tile source={new ol.source.Stamen({ layer: 'watercolor' })}/>
+            <layer.Tile source={new olSourceStamen({ layer: 'watercolor' })}/>
           </Layers>
           <Overlays>
             <Overlay ref={comp => this.overlayComp = comp}>
@@ -44,7 +45,7 @@ export class AppOverlay extends React.Component<any,any> {
         <pre>{`
         <Map onClick={this.showPopup}>
           <Layers>
-            <layer.Tile source={new ol.source.Stamen({ layer: 'watercolor' })}/>
+            <layer.Tile source={new olSourceStamen({ layer: 'watercolor' })}/>
           </Layers>
           <Overlays>
             <Overlay ref={comp => this.overlayComp = comp}>
