@@ -1,46 +1,45 @@
-import * as React from 'react';
-import * as ol from 'openlayers';
-import {Util} from '../util';
-import {Map} from '../map';
+import * as React from "react"
+import olControlAttribution from "ol/control/attribution"
+import { Util } from "../util"
+import { MapContext, Map } from "../map"
 
-export class Attribution extends React.Component<any, any> {
+class Attribution extends React.Component<any, any> {
+    control: olControlAttribution
 
-  control: ol.control.Attribution;
-
-  options: any = {
-    className: undefined,
-    target: undefined,
-    collapsible: undefined,
-    collapsed: undefined,
-    tipLabel: undefined,
-    label: undefined,
-    collapseLabel: undefined,
-    render: undefined
-  };
-
-  events: any = {
-    'change': undefined,
-    'propertychange': undefined
-  };
-
-  constructor(props) { super(props); }
-
-  render() { return null; }
-
-  componentDidMount () {
-    let options = Util.getOptions(Object['assign'](this.options, this.props));
-    this.control = new ol.control.Attribution(options);
-    this.context.mapComp.controls.push(this.control)
-
-    let olEvents = Util.getEvents(this.events, this.props);
-    for(let eventName in olEvents) {
-      this.control.on(eventName, olEvents[eventName]);
+    options: any = {
+        className: undefined,
+        target: undefined,
+        collapsible: undefined,
+        collapsed: undefined,
+        tipLabel: undefined,
+        label: undefined,
+        collapseLabel: undefined,
+        render: undefined,
     }
-  }
 
+    events: any = {
+        change: undefined,
+        propertychange: undefined,
+    }
+
+    constructor(props) {
+        super(props)
+    }
+
+    render() {
+        return null
+    }
+
+    componentDidMount() {
+        let options = Util.getOptions(Object["assign"](this.options, this.props))
+        this.control = new olControlAttribution(options)
+        this.props.mapComp.controls.push(this.control)
+
+        let olEvents = Util.getEvents(this.events, this.props)
+        for (let eventName in olEvents) {
+            this.control.on(eventName, olEvents[eventName])
+        }
+    }
 }
 
-Attribution['contextTypes'] = {
-  mapComp: React.PropTypes.instanceOf(Map),
-  map: React.PropTypes.instanceOf(ol.Map)
-};
+export default props => <MapContext.Consumer>{mapComp => <Attribution {...props} mapComp={mapComp} />}</MapContext.Consumer>
