@@ -1,45 +1,44 @@
-import * as React from 'react';
-import * as ol from 'openlayers';
-import {Util} from '../util';
-import {Map} from '../map';
+import * as React from "react"
+import olControlFullScreen from "ol/control/fullscreen"
+import { Util } from "../util"
+import { MapContext, Map } from "../map"
 
-export class FullScreen extends React.Component<any, any> {
+class FullScreen extends React.Component<any, any> {
+    control: olControlFullScreen
 
-  control: ol.control.FullScreen;
-
-  options: any = {
-    className: undefined,
-    label: undefined,
-    labelActive: undefined,
-    tipLabel: undefined,
-    keys: undefined,
-    target: undefined,
-    source: undefined
-  };
-
-  events: any = {
-    'change': undefined,
-    'propertychange': undefined
-  };
-
-  constructor(props) { super(props); }
-
-  render() { return null; }
-
-  componentDidMount () {
-    let options = Util.getOptions(Object['assign'](this.options, this.props));
-    this.control = new ol.control.FullScreen(options);
-    this.context.mapComp.controls.push(this.control)
-
-    let olEvents = Util.getEvents(this.events, this.props);
-    for(let eventName in olEvents) {
-      this.control.on(eventName, olEvents[eventName]);
+    options: any = {
+        className: undefined,
+        label: undefined,
+        labelActive: undefined,
+        tipLabel: undefined,
+        keys: undefined,
+        target: undefined,
+        source: undefined,
     }
-  }
 
+    events: any = {
+        change: undefined,
+        propertychange: undefined,
+    }
+
+    constructor(props) {
+        super(props)
+    }
+
+    render() {
+        return null
+    }
+
+    componentDidMount() {
+        let options = Util.getOptions(Object["assign"](this.options, this.props))
+        this.control = new olControlFullScreen(options)
+        this.props.mapComp.controls.push(this.control)
+
+        let olEvents = Util.getEvents(this.events, this.props)
+        for (let eventName in olEvents) {
+            this.control.on(eventName, olEvents[eventName])
+        }
+    }
 }
 
-FullScreen['contextTypes'] = {
-  mapComp: React.PropTypes.instanceOf(Map),
-  map: React.PropTypes.instanceOf(ol.Map)
-};
+export default props => <MapContext.Consumer>{mapComp => <FullScreen {...props} mapComp={mapComp} />}</MapContext.Consumer>
