@@ -1,43 +1,42 @@
-import * as React from 'react';
-import * as ol from 'openlayers';
-import {Util} from '../util';
-import {Map} from '../map';
+import * as React from "react"
+import olControlZoomSlider from "ol/control/zoomslider"
+import { Util } from "../util"
+import { MapContext, Map } from "../map"
 
-export class ZoomSlider extends React.Component<any, any> {
+class ZoomSlider extends React.Component<any, any> {
+    control: olControlZoomSlider
 
-  control: ol.control.ZoomSlider;
-
-  options: any = {
-    duration: undefined,
-    className: undefined,
-    maxResolution: undefined,
-    minResolution: undefined,
-    render: undefined
-  };
-
-  events: any = {
-    'change': undefined,
-    'propertychange': undefined
-  };
-
-  constructor(props) { super(props); }
-
-  render() { return null; }
-
-  componentDidMount () {
-    let options = Util.getOptions(Object['assign'](this.options, this.props));
-    this.control = new ol.control.ZoomSlider(options);
-    this.context.mapComp.controls.push(this.control)
-
-    let olEvents = Util.getEvents(this.events, this.props);
-    for(let eventName in olEvents) {
-      this.control.on(eventName, olEvents[eventName]);
+    options: any = {
+        duration: undefined,
+        className: undefined,
+        maxResolution: undefined,
+        minResolution: undefined,
+        render: undefined,
     }
-  }
 
+    events: any = {
+        change: undefined,
+        propertychange: undefined,
+    }
+
+    constructor(props) {
+        super(props)
+    }
+
+    render() {
+        return null
+    }
+
+    componentDidMount() {
+        let options = Util.getOptions(Object["assign"](this.options, this.props))
+        this.control = new olControlZoomSlider(options)
+        this.props.mapComp.controls.push(this.control)
+
+        let olEvents = Util.getEvents(this.events, this.props)
+        for (let eventName in olEvents) {
+            this.control.on(eventName, olEvents[eventName])
+        }
+    }
 }
 
-ZoomSlider['contextTypes'] = {
-  mapComp: React.PropTypes.instanceOf(Map),
-  map: React.PropTypes.instanceOf(ol.Map)
-};
+export default props => <MapContext.Consumer>{mapComp => <ZoomSlider {...props} mapComp={mapComp} />}</MapContext.Consumer>
