@@ -1,49 +1,44 @@
-import * as React from 'react';
-import * as ol from 'openlayers';
-import {Util} from '../util';
-import {Map} from '../map';
+import * as React from "react"
+import olControlZoom from "ol/control/zoom"
+import { Util } from "../util"
+import { MapContext, Map } from "../map"
 
-export class Zoom extends React.Component<any, any> {
+class Zoom extends React.Component<any, any> {
+    control: olControlZoom
 
-  control: ol.control.Zoom;
-
-  options: any = {
-    duration: undefined,
-    className: undefined,
-    zoomInLabel: undefined,
-    zoomOutLabel: undefined,
-    zoomInTipLabel: undefined,
-    zoomOutTipLabel: undefined,
-    delta: undefined
-  };
-
-  events: any = {
-    'change': undefined,
-    'propertychange': undefined
-  };
-
-  constructor(props) {
-    super(props);
-  }
-
-  render() {
-    return null;
-  }
-
-  componentDidMount () {
-    let options = Util.getOptions(Object['assign'](this.options, this.props));
-    this.control = new ol.control.Zoom(options);
-    this.context.mapComp.controls.push(this.control)
-
-    let olEvents = Util.getEvents(this.events, this.props);
-    for(let eventName in olEvents) {
-      this.control.on(eventName, olEvents[eventName]);
+    options: any = {
+        duration: undefined,
+        className: undefined,
+        zoomInLabel: undefined,
+        zoomOutLabel: undefined,
+        zoomInTipLabel: undefined,
+        zoomOutTipLabel: undefined,
+        delta: undefined,
     }
-  }
 
+    events: any = {
+        change: undefined,
+        propertychange: undefined,
+    }
+
+    constructor(props) {
+        super(props)
+    }
+
+    render() {
+        return null
+    }
+
+    componentDidMount() {
+        let options = Util.getOptions(Object["assign"](this.options, this.props))
+        this.control = new olControlZoom(options)
+        this.props.mapComp.controls.push(this.control)
+
+        let olEvents = Util.getEvents(this.events, this.props)
+        for (let eventName in olEvents) {
+            this.control.on(eventName, olEvents[eventName])
+        }
+    }
 }
 
-Zoom['contextTypes'] = {
-  mapComp: React.PropTypes.instanceOf(Map),
-  map: React.PropTypes.instanceOf(ol.Map)
-};
+export default props => <MapContext.Consumer>{mapComp => <Zoom {...props} mapComp={mapComp} />}</MapContext.Consumer>
