@@ -1,22 +1,15 @@
-import {Popup} from './popup';
-import {GoogleStreetViewPanorama} from './google-street-view-panorama';
-import {ClusterStyle} from './style/cluster-style';
-import {MarkerStyle} from './style/marker-style';
-import {GeoCoderControl} from './control/geo-coder-control';
-import {GeoCoderComponent} from './control/geo-coder-component';
+import { Popup } from "./popup"
+import { GoogleStreetViewPanorama } from "./google-street-view-panorama"
+import { ClusterStyle } from "./style/cluster-style"
+import { MarkerStyle } from "./style/marker-style"
 
 let custom = {
-  style: {
-    MarkerStyle: MarkerStyle,
-    ClusterStyle: ClusterStyle
-  },
-  control: {
-    GeoCoderControl: GeoCoderControl,
-    GeoCoderComponent: GeoCoderComponent
-  },
-  Popup: Popup,
-  GoogleStreetViewPanorama: GoogleStreetViewPanorama
-};
+    style: {
+        MarkerStyle: MarkerStyle,
+        ClusterStyle: ClusterStyle,
+    },
+    Popup: Popup,
+    GoogleStreetViewPanorama: GoogleStreetViewPanorama,
+}
 
-export {custom};
-
+export { custom }
