@@ -1,44 +1,39 @@
-import * as React from 'react';
-import * as ol from 'openlayers';
+import * as React from "react"
 
-import './popup.css';
+import "./popup.css"
 
 export class Popup extends React.Component<any, any> {
-  containerEl: HTMLElement;
-  contentEl: HTMLElement;
-  contentClose:HTMLElement;
+    containerEl: HTMLElement
+    contentEl: HTMLElement
+    contentClose: HTMLElement
 
-  constructor(props) {
-    console.log(2222222222);
-    super(props);
-  }
+    constructor(props) {
+        console.log(2222222222)
+        super(props)
+    }
 
-  componentDidMount(){
-    this.contentClose.addEventListener("click",()=>{
-      this.containerEl.style.display='none';
-    });
-  }
+    componentDidMount() {
+        this.contentClose.addEventListener("click", () => {
+            this.containerEl.style.display = "none"
+        })
+    }
 
-  render() {
-    console.log(3333333333);
-    return (
-      <div className="olPopup" ref={el => this.containerEl = el}>
-        <a className="olPopupCloser"
-          href="javascript:void(0)"
-          ref={el => this.contentClose = el}
-        ></a>
-        <div className="olPopupContents" ref={el => this.contentEl = el}></div>
-      </div>
-    );
-  }
+    render() {
+        console.log(3333333333)
+        return (
+            <div className="olPopup" ref={el => (this.containerEl = el)}>
+                <a className="olPopupCloser" href="javascript:void(0)" ref={el => (this.contentClose = el)} />
+                <div className="olPopupContents" ref={el => (this.contentEl = el)} />
+            </div>
+        )
+    }
 
-  setContents(html) {
-    this.contentEl.innerHTML = html;
-  }
+    setContents(html) {
+        this.contentEl.innerHTML = html
+    }
 
-  show(bottomDistance: string = '12px') {
-    this.containerEl.style.bottom = bottomDistance;
-    this.containerEl.style.display = 'block';
-  }
-
+    show(bottomDistance: string = "12px") {
+        this.containerEl.style.bottom = bottomDistance
+        this.containerEl.style.display = "block"
+    }
 }
