@@ -1,27 +1,30 @@
-import * as React from "react";
-import * as ReactDOM from "react-dom";
-import * as ol from 'openlayers';
+import * as React from "react"
+import * as ReactDOM from "react-dom"
+import olStyleStyle from "ol/style/style"
+import olStyleIcon from "ol/style/icon"
 
 export class MarkerStyle {
-  src: string = 'https://openlayers.org/en/v4.0.1/examples/data/icon.png';
+    src: string = "https://openlayers.org/en/v4.0.1/examples/data/icon.png"
 
-  constructor(src?: string) {
-    this.src = src;
-  }
+    constructor(src?: string) {
+        this.src = src
+    }
 
-  style = new ol.style.Style({
-    image: new ol.style.Icon(/** @type {olx.style.IconOptions} */ ({
-      src: this.src
-    }))
-  });
-
-  selectStyleFunction = (feature)  => {
-    return new ol.style.Style({
-      image: new ol.style.Icon({
-          anchor: [0.5, 0.96],
-          color: '#4271AE',
-          src: 'https://openlayers.org/en/v4.0.1/examples/data/dot.png'
-        })
+    style = new olStyleStyle({
+        image: new olStyleIcon(
+            /** @type {olx.style.IconOptions} */ ({
+                src: this.src,
+            })
+        ),
     })
-  };
+
+    selectStyleFunction = feature => {
+        return new olStyleStyle({
+            image: new olStyleIcon({
+                anchor: [0.5, 0.96],
+                color: "#4271AE",
+                src: "https://openlayers.org/en/v4.0.1/examples/data/dot.png",
+            }),
+        })
+    }
 }
