@@ -1,29 +1,26 @@
-import {Controls, control} from './controls/index';
-import {Interactions, interaction} from './interactions/index';
-import {Layers, layer} from './layers/index';
-import {Overlays} from './overlays/overlays';
-import {custom} from './custom/index';
+import { Controls, control } from "./controls/index"
+import { Interactions, interaction } from "./interactions/index"
+import { Layers, layer } from "./layers/index"
+import { Overlays } from "./overlays/overlays"
+import { custom } from "./custom/index"
 
-import {Map} from  './map';
-import {Overlay} from './overlay';
-import {Util} from './util';
+import { Map } from "./map"
+import { default as Overlay } from "./overlay"
+import { Util } from "./util"
 
 export {
-  //groups
-  Interactions,
-  Layers,
-  Overlays,
-  Controls,
-
-  //name spaces
-  layer,
-  custom,
-  control,
-  interaction,
-
-  //Objects
-  Map,
-  Overlay,
-
-  Util
-};
+    //groups
+    Interactions,
+    Layers,
+    Overlays,
+    Controls,
+    //name spaces
+    layer,
+    custom,
+    control,
+    interaction,
+    //Objects
+    Map,
+    Overlay,
+    Util,
+}
