@@ -1,59 +1,58 @@
-import * as React from 'react';
-import * as ol from 'openlayers';
-import {Util} from "../util";
-import {Map} from '../map';
+import * as React from "react"
+import olInteractionDoubleClickZoom from "ol/interaction/doubleclickzoom"
+import { Util } from "../util"
+import { MapContext, Map } from "../map"
 
-export class DoubleClickZoom extends React.Component<any, any> {
+class DoubleClickZoom extends React.Component<any, any> {
+    interaction: olInteractionDoubleClickZoom
 
-  interaction: ol.interaction.DoubleClickZoom;
-
-  options: any = {
-    duration: undefined,
-    delta: undefined
-  };
-
-  events: any = {
-    'change': undefined,
-    'change:active': undefined,
-    'propertychange': undefined
-  };
-
-  constructor(props) { super(props); }
-
-  render() { return null; }
-
-  componentDidMount () {
-    let options = Util.getOptions(Object['assign'](this.options, this.props));
-    this.interaction = new ol.interaction.DoubleClickZoom(options);
-    this.context.mapComp.interactions.push(this.interaction)
-
-    let olEvents = Util.getEvents(this.events, this.props);
-    for(let eventName in olEvents) {
-      this.interaction.on(eventName, olEvents[eventName]);
+    options: any = {
+        duration: undefined,
+        delta: undefined,
     }
-  }
 
-  componentWillReceiveProps (nextProps) {
-    if(nextProps !== this.props){
-      this.context.mapComp.map.removeInteraction(this.interaction);
-      let options = Util.getOptions(Object['assign'](this.options, nextProps));
-      this.interaction = new ol.interaction.DoubleClickZoom(options);
-      this.context.mapComp.map.addInteraction(this.interaction);
-
-      let olEvents = Util.getEvents(this.events, this.props);
-      for(let eventName in olEvents) {
-        this.interaction.on(eventName, olEvents[eventName]);
-      }
+    events: any = {
+        change: undefined,
+        "change:active": undefined,
+        propertychange: undefined,
     }
-  }
-  
-  componentWillUnmount () {
-    this.context.mapComp.map.removeInteraction(this.interaction);
-  }
 
+    constructor(props) {
+        super(props)
+    }
+
+    render() {
+        return null
+    }
+
+    componentDidMount() {
+        let options = Util.getOptions(Object["assign"](this.options, this.props))
+        this.interaction = new olInteractionDoubleClickZoom(options)
+        this.props.mapComp.interactions.push(this.interaction)
+
+        let olEvents = Util.getEvents(this.events, this.props)
+        for (let eventName in olEvents) {
+            this.interaction.on(eventName, olEvents[eventName])
+        }
+    }
+
+    componentWillReceiveProps(nextProps) {
+        if (nextProps !== this.props) {
+            this.props.mapComp.map.removeInteraction(this.interaction)
+            let options = Util.getOptions(Object["assign"](this.options, nextProps))
+            this.interaction = new olInteractionDoubleClickZoom(options)
+            this.props.mapComp.map.addInteraction(this.interaction)
+
+            let olEvents = Util.getEvents(this.events, this.props)
+            for (let eventName in olEvents) {
+                this.interaction.on(eventName, olEvents[eventName])
+            }
+        }
+    }
+
+    componentWillUnmount() {
+        this.props.mapComp.map.removeInteraction(this.interaction)
+    }
 }
 
-DoubleClickZoom['contextTypes'] = {
-  mapComp: React.PropTypes.instanceOf(Map),
-  map: React.PropTypes.instanceOf(ol.Map)
-};
+export default props => <MapContext.Consumer>{mapComp => <DoubleClickZoom {...props} mapComp={mapComp} />}</MapContext.Consumer>
