@@ -1,78 +1,77 @@
-import * as React from 'react';
-import * as ol from 'openlayers';
-import {Util} from "../util";
-import {Map} from '../map';
+import * as React from "react"
+import olInteractionSelect from "ol/interaction/select"
+import { Util } from "../util"
+import { MapContext, Map } from "../map"
 
-export class Select extends React.Component<any, any> {
+class Select extends React.Component<any, any> {
+    interaction: olInteractionSelect
 
-  interaction: ol.interaction.Select;
-
-  options: any = {
-    addCondition: undefined,
-    condition: undefined,
-    layers: undefined,
-    style: undefined,
-    removeCondition: undefined,
-    toggleCondition: undefined,
-    multi: undefined,
-    features: undefined,
-    filter: undefined,
-    wrapX: undefined,
-    hitTolerance: undefined
-  };
-
-  events: any = {
-    'change': undefined,
-    'change:active': undefined,
-    'propertychange': undefined,
-    'select': undefined
-  };
-
-  constructor(props) { super(props); }
-
-  render() { return null; }
-
-  componentDidMount () {
-    if (this.props.instance) {
-      this.interaction = this.props.instance;
-    } else {
-      let options = Util.getOptions(Object['assign'](this.options, this.props));
-      this.interaction = new ol.interaction.Select(options);
+    options: any = {
+        addCondition: undefined,
+        condition: undefined,
+        layers: undefined,
+        style: undefined,
+        removeCondition: undefined,
+        toggleCondition: undefined,
+        multi: undefined,
+        features: undefined,
+        filter: undefined,
+        wrapX: undefined,
+        hitTolerance: undefined,
     }
-    this.context.mapComp.interactions.push(this.interaction)
-    
-    let olEvents = Util.getEvents(this.events, this.props);
-    for(let eventName in olEvents) {
-      this.interaction.on(eventName, olEvents[eventName]);
+
+    events: any = {
+        change: undefined,
+        "change:active": undefined,
+        propertychange: undefined,
+        select: undefined,
     }
-  }
 
-  componentWillReceiveProps (nextProps) {
-    if(nextProps !== this.props){
-      this.context.mapComp.map.removeInteraction(this.interaction);
-
-      if (this.props.instance) {
-        this.interaction = this.props.instance;
-      } else {
-        let options = Util.getOptions(Object['assign'](this.options, nextProps));
-        this.interaction = new ol.interaction.Select(options);
-      }
-      this.context.mapComp.map.addInteraction(this.interaction);
-
-      let olEvents = Util.getEvents(this.events, this.props);
-      for(let eventName in olEvents) {
-        this.interaction.on(eventName, olEvents[eventName]);
-      }
+    constructor(props) {
+        super(props)
     }
-  }
-  
-  componentWillUnmount () {
-    this.context.mapComp.map.removeInteraction(this.interaction);
-  }
 
+    render() {
+        return null
+    }
+
+    componentDidMount() {
+        if (this.props.instance) {
+            this.interaction = this.props.instance
+        } else {
+            let options = Util.getOptions(Object["assign"](this.options, this.props))
+            this.interaction = new olInteractionSelect(options)
+        }
+        this.props.mapComp.interactions.push(this.interaction)
+
+        let olEvents = Util.getEvents(this.events, this.props)
+        for (let eventName in olEvents) {
+            this.interaction.on(eventName, olEvents[eventName])
+        }
+    }
+
+    componentWillReceiveProps(nextProps) {
+        if (nextProps !== this.props) {
+            this.props.mapComp.map.removeInteraction(this.interaction)
+
+            if (this.props.instance) {
+                this.interaction = this.props.instance
+            } else {
+                let options = Util.getOptions(Object["assign"](this.options, nextProps))
+                this.interaction = new olInteractionSelect(options)
+            }
+            this.props.mapComp.map.addInteraction(this.interaction)
+
+            let olEvents = Util.getEvents(this.events, this.props)
+            for (let eventName in olEvents) {
+                this.interaction.on(eventName, olEvents[eventName])
+            }
+        }
+    }
+
+    componentWillUnmount() {
+        this.props.mapComp.map.removeInteraction(this.interaction)
+    }
 }
 
-Select['contextTypes'] = {
-  mapComp: React.PropTypes.instanceOf(Map),
-  map: React.PropTypes.instanceOf(ol.Map)
-};
+export default props => <MapContext.Consumer>{mapComp => <Select {...props} mapComp={mapComp} />}</MapContext.Consumer>
