@@ -1,7 +1,7 @@
-import * as React from 'react';
+import * as React from "react"
 
 export class Overlays extends React.Component<any, any> {
-  render() {
-    return (<div>{this.props.children}</div>);
-  }
+    render() {
+        return <div>{this.props.children}</div>
+    }
 }
